Use absolute path for 'find'
Now when the router boots and the tincd being ran from the WAN Up
script, the custom routes somehow doesn't get read. I suspect it's
something to do with the 'find' utility residing in /opt or /jffs
affects the function of reading the file list.

This will force the script to use the 'find' utility provided by the
busybox embedded in ROM.